内容中心-图片公告

+++ helper-custom-server/src/main/java/com/qc/printers/custom/user/service/impl/RoleServiceImpl.java
@@ -153,12 +153,18 @@ public class RoleServiceImpl implements RoleService {
             for (String nid :
                     ids
             ) {
+                if (nid.equals("1")) {
+                    throw new CustomException("系统角色不可删除");
+                }
                 iSysRoleService.removeById(Long.valueOf(nid));
                 LambdaQueryWrapper<SysRoleMenu> sysRoleMenuLambdaQueryWrapper = new LambdaQueryWrapper<>();
                 sysRoleMenuLambdaQueryWrapper.eq(SysRoleMenu::getRoleId, nid);
                 iSysRoleMenuService.remove(sysRoleMenuLambdaQueryWrapper);
             }
         } else {
+            if (id.equals("1")) {
+                throw new CustomException("系统角色不可删除");
+            }
             iSysRoleService.removeById(Long.valueOf(id));
             LambdaQueryWrapper<SysRoleMenu> sysRoleMenuLambdaQueryWrapper = new LambdaQueryWrapper<>();
             sysRoleMenuLambdaQueryWrapper.eq(SysRoleMenu::getRoleId, id);
